requirements.txt
asgiref==3.3.4
Django==3.2
django-recaptcha==2.0.6
django-render-partial==0.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
